Archive (1)/requirements.txt
fastapi
uvicorn
pillow
pytesseract
requests
sqlalchemy